meals/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
inflect==6.0.2
num2words==0.5.12
Pillow==9.4.0
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
quantulum3==0.8.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
